Interrupt requests
Interrupt requests have now been added to the kernel. It is responding
to these as expected.

+++ include/types.h
@@ -20,14 +20,14 @@ typedef          long ssize_t;
 #endif
 
 /* Standard integer types */
-typedef unsigned char  u8;
-typedef          char  s8;
-typedef unsigned short u16;
-typedef          short s16;
-typedef unsigned int   u32;
-typedef          int   s32;
-typedef unsigned long  u64;
-typedef          long  s64;
+typedef unsigned char  		u8;
+typedef          char  		s8;
+typedef unsigned short 		u16;
+typedef          short 		s16;
+typedef unsigned int   		u32;
+typedef          int   		s32;
+typedef unsigned long long 	u64;
+typedef          long long 	s64;
 
 #ifdef __64BIT__
 typedef unsigned long long uintmax_t;
